pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.3.0",
    "elasticsearch>=9.2.0",
    "fastapi-mcp>=0.4.0",
    "fastapi[standard]>=0.120.4",
    "flagembedding>=1.3.5",
    "google-api-core>=2.28.1",
    "google-api-python-client>=2.186.0",
    "google-auth>=2.42.1",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-core>=2.5.0",
    "groq>=0.33.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-google-community>=3.0.0",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-ollama>=1.0.0",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.2",
    "openai>=2.6.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "wikipedia>=1.4.0",
]
